Add symbolic link and reference to fusion-menu header
Note that we don't use a submodule here for the sake of simplicity; we
instead just require the sibling project to be checked out alongside
this one (which will usually be the case because the different plug-ins
will live alongside each other as submodules in the parent project).

Also note how this interacts with the "-undefined dynamic_lookup" linker
option: we don't have any real (strong) dependencies between the
different plug-ins seeing as they don't have to link against each other,
and that in turn means we can get very fast build times (ie. a single
change to one plug-in will not necessarily trigger a rebuild for all the
others as well).

The only hard requirement here is that the Fusion framework headers
already be available in the build products directory by the time we
compile the plug-in. In practice this will usually be the case because
the superproject will compile the framework first. If I find that it is
not working out that way for whatever reason, I can always add another
symbolic link.

Signed-off-by: Wincent Colaiuta <win@wincent.com>